+++ detection/utils.py
@@ -1,18 +1,15 @@
 import torch
-import numpy as np
 from config.classes.coco_labels import CLASS_NAMES
 
 
 def frame_to_tensor(frame):
     """
-    convert frame to tensor
-    :param frame: frame of video
+    Convert image frame to tensor
+    :param frame: img
     :return: 3R tensor
     """
     frame = torch.from_numpy(frame).float() / 255.0
     frame = frame.permute(2, 0, 1)
-    # TODO: release batchs
-    # FIXME: possible need return list of frame
     return frame
 
 
@@ -21,10 +18,9 @@ def collate_fn(batch):
 
 
 def class_names():
-    """:return class names"""
+    """
+    Get coco class names
+    :return dictionary {id : name}"""
     return CLASS_NAMES
 
 
-def color_bounding_box(classes):
-    """ :return array of colors"""
-    return np.random.uniform(0, 255, size=(len(classes), 3))
